BACKEND/requirements.txt
django==6.0.0
djangorestframework==3.15.2
django-cors-headers==4.3.1
twilio==9.2.3
pywebpush==2.0.0

